+++ frontend/src/paginas/EditarCliente/EditarCliente.js
@@ -2,8 +2,8 @@ import { Box, Button, Grid, Stack } from "@mui/material";
 import { useState } from "react";
 import { useLocation } from "react-router-dom";
 import { atualizarCliente, removerCliente } from "../../api/api";
-import CartaoCliente from "./CartaoCliente";
-import CartaoContato from "./CartaoContato";
+import CartaoCliente from "./cartoes/CartaoCliente";
+import CartaoContato from "./cartoes/CartaoContato";
 import ModalEditarCliente from "./modais/ModalEditarCliente";
 import ModalEditarContato from "./modais/ModalEditarContato";
 import ModalAdicionarContato from "./modais/ModalAdicionarContato";
